Corrections for performance workbench

+++ open-metadata-conformance-suite/open-metadata-conformance-suite-server/src/main/java/org/odpi/openmetadata/conformance/tests/performance/purge/TestEntityPurge.java
@@ -28,8 +28,11 @@ public class TestEntityPurge extends OpenMetadataPerformanceTestCase
     private static final String A_FIND_ENTITIES        = TEST_CASE_ID + "-findEntitiesByProperty";
     private static final String A_FIND_ENTITIES_MSG    = "Repository performs search for unordered first instancesPerType deleted instances of type: ";
 
-    private static final String A_FIND_ENTITIES_RC     = TEST_CASE_ID + "-findEntitiesByProperty";
+    private static final String A_FIND_ENTITIES_RC     = TEST_CASE_ID + "-findEntitiesByProperty-rc";
     private static final String A_FIND_ENTITIES_RC_MSG = "Repository performs search for unordered first instancesPerType reference copy instances of type: ";
+
+    private static final String A_DELETE     = TEST_CASE_ID + "-deleteEntity";
+    private static final String A_DELETE_MSG = "Repository delete of instances of type: ";
 
     private static final String A_PURGE     = TEST_CASE_ID + "-purgeEntity";
     private static final String A_PURGE_MSG = "Repository purge of instances of type: ";
@@ -71,6 +74,7 @@ public class TestEntityPurge extends OpenMetadataPerformanceTestCase
         int numInstances = super.getInstancesPerType();
 
         Set<String> keys = getKeys(metadataCollection, numInstances);
+        deleteEntities(metadataCollection, keys);
         purgeEntities(metadataCollection, keys);
         Set<String> rcKeys = getReferenceCopyKeys(metadataCollection, numInstances);
         purgeEntityReferenceCopies(metadataCollection, rcKeys);
@@ -79,37 +83,45 @@ public class TestEntityPurge extends OpenMetadataPerformanceTestCase
     }
 
     /**
-     * Retrieve a list of entities that are deleted.
+     * Retrieve a list of entities that are homed in the technology under test's repository.
      * @param metadataCollection through which to call findEntitiesByProperty
      * @param numInstances number of instances to retrieve
-     * @return set of GUIDs that are deleted
+     * @return set of GUIDs that are homed in the technology under test's repository
      * @throws Exception on any errors
      */
     private Set<String> getKeys(OMRSMetadataCollection metadataCollection, int numInstances) throws Exception
     {
-        List<InstanceStatus> deleted = new ArrayList<>();
-        deleted.add(InstanceStatus.DELETED);
+        final String methodName = "getKeys";
+        OMRSRepositoryHelper repositoryHelper = super.getRepositoryHelper();
+        InstanceProperties byMetadataCollectionId = repositoryHelper.addStringPropertyToInstance(testCaseId,
+                null,
+                "metadataCollectionId",
+                repositoryHelper.getExactMatchRegex(performanceWorkPad.getTutMetadataCollectionId()),
+                methodName);
         long start = System.nanoTime();
         List<EntityDetail> entities = metadataCollection.findEntitiesByProperty(workPad.getLocalServerUserId(),
                 entityDef.getGUID(),
-                null,
-                null,
+                byMetadataCollectionId,
+                MatchCriteria.ALL,
                 0,
-                deleted,
+                null,
                 null,
                 null,
                 null,
                 null,
                 numInstances);
         long elapsedTime = (System.nanoTime() - start) / 1000000;
-        assertCondition(entities != null,
-                A_FIND_ENTITIES,
-                A_FIND_ENTITIES_MSG + testTypeName,
-                PerformanceProfile.ENTITY_SEARCH.getProfileId(),
-                null,
-                "findEntitiesByProperty",
-                elapsedTime);
-        return entities == null ? null : entities.stream().map(EntityDetail::getGUID).collect(Collectors.toSet());
+        if (entities != null) {
+            assertCondition(true,
+                    A_FIND_ENTITIES,
+                    A_FIND_ENTITIES_MSG + testTypeName,
+                    PerformanceProfile.ENTITY_SEARCH.getProfileId(),
+                    null,
+                    "findEntitiesByProperty",
+                    elapsedTime);
+            return entities.stream().map(EntityDetail::getGUID).collect(Collectors.toSet());
+        }
+        return null;
     }
 
     /**
@@ -141,14 +153,61 @@ public class TestEntityPurge extends OpenMetadataPerformanceTestCase
                 null,
                 numInstances);
         long elapsedTime = (System.nanoTime() - start) / 1000000;
-        assertCondition(entities != null,
-                A_FIND_ENTITIES_RC,
-                A_FIND_ENTITIES_RC_MSG + testTypeName,
-                PerformanceProfile.ENTITY_SEARCH.getProfileId(),
-                null,
-                "findEntitiesByProperty",
-                elapsedTime);
-        return entities == null ? null : entities.stream().map(EntityDetail::getGUID).collect(Collectors.toSet());
+        if (entities != null) {
+            assertCondition(true,
+                    A_FIND_ENTITIES_RC,
+                    A_FIND_ENTITIES_RC_MSG + testTypeName,
+                    PerformanceProfile.ENTITY_SEARCH.getProfileId(),
+                    null,
+                    "findEntitiesByProperty",
+                    elapsedTime);
+            return entities.stream().map(EntityDetail::getGUID).collect(Collectors.toSet());
+        }
+        return null;
+    }
+
+    /**
+     * Attempt to delete a number of existing entities.
+     * @param metadataCollection through which to call deleteEntity
+     * @param keys GUIDs of entities to delete
+     * @throws Exception on any errors
+     */
+    private void deleteEntities(OMRSMetadataCollection metadataCollection, Set<String> keys) throws Exception
+    {
+
+        final String methodName = "deleteEntity";
+
+        if (keys != null) {
+            try {
+                for (String guid : keys) {
+                    long start = System.nanoTime();
+                    EntityDetail result = metadataCollection.deleteEntity(workPad.getLocalServerUserId(),
+                            entityDef.getGUID(),
+                            entityDef.getName(),
+                            guid);
+                    long elapsedTime = (System.nanoTime() - start) / 1000000;
+                    assertCondition(result != null,
+                            A_DELETE,
+                            A_DELETE_MSG + testTypeName,
+                            PerformanceProfile.ENTITY_DELETE.getProfileId(),
+                            null,
+                            methodName,
+                            elapsedTime);
+                }
+            } catch (FunctionNotSupportedException exception) {
+                super.addNotSupportedAssertion(A_DELETE,
+                        A_DELETE_MSG + testTypeName,
+                        PerformanceProfile.ENTITY_DELETE.getProfileId(),
+                        null);
+            } catch (Exception exc) {
+                String operationDescription = "delete entity of type " + entityDef.getName();
+                Map<String, String> parameters = new HashMap<>();
+                parameters.put("typeGUID", entityDef.getGUID());
+                String msg = this.buildExceptionMessage(testCaseId, methodName, operationDescription, parameters, exc.getClass().getSimpleName(), exc.getMessage());
+                throw new Exception(msg, exc);
+            }
+        }
+
     }
 
     /**
@@ -162,34 +221,35 @@ public class TestEntityPurge extends OpenMetadataPerformanceTestCase
 
         final String methodName = "purgeEntity";
 
-        try {
-            for (String guid : keys) {
-                long start = System.nanoTime();
-                metadataCollection.purgeEntity(workPad.getLocalServerUserId(),
-                        entityDef.getGUID(),
-                        entityDef.getName(),
-                        guid);
-                long elapsedTime = (System.nanoTime() - start) / 1000000;
-                assertCondition(true,
-                        A_PURGE,
+        if (keys != null) {
+            try {
+                for (String guid : keys) {
+                    long start = System.nanoTime();
+                    metadataCollection.purgeEntity(workPad.getLocalServerUserId(),
+                            entityDef.getGUID(),
+                            entityDef.getName(),
+                            guid);
+                    long elapsedTime = (System.nanoTime() - start) / 1000000;
+                    assertCondition(true,
+                            A_PURGE,
+                            A_PURGE_MSG + testTypeName,
+                            PerformanceProfile.ENTITY_PURGE.getProfileId(),
+                            null,
+                            methodName,
+                            elapsedTime);
+                }
+            } catch (FunctionNotSupportedException exception) {
+                super.addNotSupportedAssertion(A_PURGE,
                         A_PURGE_MSG + testTypeName,
                         PerformanceProfile.ENTITY_PURGE.getProfileId(),
-                        null,
-                        methodName,
-                        elapsedTime);
+                        null);
+            } catch (Exception exc) {
+                String operationDescription = "purge entity of type " + entityDef.getName();
+                Map<String, String> parameters = new HashMap<>();
+                parameters.put("typeGUID", entityDef.getGUID());
+                String msg = this.buildExceptionMessage(testCaseId, methodName, operationDescription, parameters, exc.getClass().getSimpleName(), exc.getMessage());
+                throw new Exception(msg, exc);
             }
-        } catch (FunctionNotSupportedException exception) {
-            super.addNotSupportedAssertion(A_PURGE,
-                    A_PURGE_MSG + testTypeName,
-                    PerformanceProfile.ENTITY_PURGE.getProfileId(),
-                    null);
-            return;
-        } catch (Exception exc) {
-            String operationDescription = "purge entity of type " + entityDef.getName();
-            Map<String, String> parameters = new HashMap<>();
-            parameters.put("typeGUID", entityDef.getGUID());
-            String msg = this.buildExceptionMessage(testCaseId, methodName, operationDescription, parameters, exc.getClass().getSimpleName(), exc.getMessage());
-            throw new Exception(msg, exc);
         }
 
     }
@@ -205,35 +265,36 @@ public class TestEntityPurge extends OpenMetadataPerformanceTestCase
 
         final String methodName = "purgeEntityReferenceCopy";
 
-        try {
-            for (String guid : keys) {
-                long start = System.nanoTime();
-                metadataCollection.purgeEntityReferenceCopy(workPad.getLocalServerUserId(),
-                        guid,
-                        entityDef.getGUID(),
-                        entityDef.getName(),
-                        performanceWorkPad.getReferenceCopyMetadataCollectionId());
-                long elapsedTime = (System.nanoTime() - start) / 1000000;
-                assertCondition(true,
-                        A_PURGE_RC,
+        if (keys != null) {
+            try {
+                for (String guid : keys) {
+                    long start = System.nanoTime();
+                    metadataCollection.purgeEntityReferenceCopy(workPad.getLocalServerUserId(),
+                            guid,
+                            entityDef.getGUID(),
+                            entityDef.getName(),
+                            performanceWorkPad.getReferenceCopyMetadataCollectionId());
+                    long elapsedTime = (System.nanoTime() - start) / 1000000;
+                    assertCondition(true,
+                            A_PURGE_RC,
+                            A_PURGE_RC_MSG + testTypeName,
+                            PerformanceProfile.ENTITY_PURGE.getProfileId(),
+                            null,
+                            methodName,
+                            elapsedTime);
+                }
+            } catch (FunctionNotSupportedException exception) {
+                super.addNotSupportedAssertion(A_PURGE_RC,
                         A_PURGE_RC_MSG + testTypeName,
                         PerformanceProfile.ENTITY_PURGE.getProfileId(),
-                        null,
-                        methodName,
-                        elapsedTime);
+                        null);
+            } catch (Exception exc) {
+                String operationDescription = "purge reference copy entity of type " + entityDef.getName();
+                Map<String, String> parameters = new HashMap<>();
+                parameters.put("typeGUID", entityDef.getGUID());
+                String msg = this.buildExceptionMessage(testCaseId, methodName, operationDescription, parameters, exc.getClass().getSimpleName(), exc.getMessage());
+                throw new Exception(msg, exc);
             }
-        } catch (FunctionNotSupportedException exception) {
-            super.addNotSupportedAssertion(A_PURGE_RC,
-                    A_PURGE_RC_MSG + testTypeName,
-                    PerformanceProfile.ENTITY_PURGE.getProfileId(),
-                    null);
-            return;
-        } catch (Exception exc) {
-            String operationDescription = "purge reference copy entity of type " + entityDef.getName();
-            Map<String, String> parameters = new HashMap<>();
-            parameters.put("typeGUID", entityDef.getGUID());
-            String msg = this.buildExceptionMessage(testCaseId, methodName, operationDescription, parameters, exc.getClass().getSimpleName(), exc.getMessage());
-            throw new Exception(msg, exc);
         }
 
     }
